# docs/swagger.yaml
basePath: /api/v1
definitions:
  firs_models.AllowanceCharge:
    properties:
      amount:
        type: number
      charge_indicator:
        type: boolean
    type: object
  firs_models.DocumentReference:
    properties:
      irn:
        type: string
      issue_date:
        type: string
    type: object
...
  firs_models.GenerateIRNRequestData:
    properties:
      invoice_number:
        type: string
    required:
    - invoice_number
    type: object
  firs_models.IRNSigningRequestData:
    properties:
      irn:
        type: string
    type: object
...
  firs_models.InvoiceDeliveryPeriod:
    properties:
      end_date:
        type: string
      start_date:
        type: string
    type: object
  firs_models.InvoiceLine:
    properties:
      discount_amount:
        type: number
      discount_rate:
        type: number
      fee_amount:
        type: number
      fee_rate:
        type: number
      hsn_code:
        type: string
      invoiced_quantity:
        type: integer
      item:
        $ref: '#/definitions/firs_models.Item'
      line_extension_amount:
        type: number
      price:
        $ref: '#/definitions/firs_models.Price'
      product_category:
        type: string
    type: object
  firs_models.InvoiceRequest:
    properties:
      _document_reference:
        items:
          $ref: '#/definitions/firs_models.DocumentReference'
        type: array
      accounting_cost:
        type: string
      accounting_customer_party:
        $ref: '#/definitions/firs_models.Party'
      accounting_supplier_party:
        $ref: '#/definitions/firs_models.Party'
      actual_delivery_date:
        type: string
      allowance_charge:
        items:
          $ref: '#/definitions/firs_models.AllowanceCharge'
        type: array
      billing_reference:
        items:
          $ref: '#/definitions/firs_models.DocumentReference'
        type: array
      business_id:
        type: string
      buyer_reference:
        type: string
      contract_document_reference:
        $ref: '#/definitions/firs_models.DocumentReference'
      dispatch_document_reference:
        $ref: '#/definitions/firs_models.DocumentReference'
      document_currency_code:
        type: string
      due_date:
        type: string
      invoice_delivery_period:
        $ref: '#/definitions/firs_models.InvoiceDeliveryPeriod'
      invoice_line:
        items:
          $ref: '#/definitions/firs_models.InvoiceLine'
        type: array
      invoice_type_code:
        type: string
      irn:
        type: string
      issue_date:
        type: string
      issue_time:
        type: string
      legal_monetary_total:
        $ref: '#/definitions/firs_models.LegalMonetaryTotal'
      note:
        type: string
      order_reference:
        type: string
      originator_document_reference:
        $ref: '#/definitions/firs_models.DocumentReference'
      payee_party:
        $ref: '#/definitions/firs_models.Party'
      payment_means:
        items:
          $ref: '#/definitions/firs_models.PaymentMeans'
        type: array
      payment_status:
        type: string
      payment_terms_note:
        type: string
      receipt_document_reference:
        $ref: '#/definitions/firs_models.DocumentReference'
      tax_currency_code:
        type: string
      tax_point_date:
        type: string
      tax_representative_party:
        $ref: '#/definitions/firs_models.Party'
      tax_total:
        items:
          $ref: '#/definitions/firs_models.TaxTotal'
        type: array
    type: object
  firs_models.Item:
    properties:
      description:
        type: string
      name:
        type: string
      sellers_item_identification:
        type: string
    type: object
  firs_models.LegalMonetaryTotal:
    properties:
      line_extension_amount:
        type: number
      payable_amount:
        type: number
      tax_exclusive_amount:
        type: number
      tax_inclusive_amount:
        type: number
    type: object
  firs_models.Party:
    properties:
      business_description:
        type: string
      email:
        type: string
      party_name:
        type: string
      postal_address:
        $ref: '#/definitions/firs_models.PostalAddress'
      telephone:
        type: string
      tin:
        type: string
    type: object
  firs_models.PaymentMeans:
    properties:
      payment_due_date:
        type: string
      payment_means_code:
        type: string
    type: object
  firs_models.PostalAddress:
    properties:
      city_name:
        type: string
      country:
        type: string
      country_code:
        type: string
      postal_zone:
        type: string
      street_name:
        type: string
    type: object
  firs_models.Price:
    properties:
      base_quantity:
        type: integer
      price_amount:
        type: number
      price_unit:
        type: string
    type: object
  firs_models.TaxCategory:
    properties:
      id:
        type: string
      percent:
        type: number
    type: object
  firs_models.TaxSubtotal:
    properties:
      tax_amount:
        type: number
      tax_category:
        $ref: '#/definitions/firs_models.TaxCategory'
      taxable_amount:
        type: number
    type: object
  firs_models.TaxTotal:
    properties:
      tax_amount:
        type: number
      tax_subtotal:
        items:
          $ref: '#/definitions/firs_models.TaxSubtotal'
        type: array
    type: object
...
  models.Response:
    properties:
      data: {}
      error: {}
      extra: {}
      message:
        type: string
      name:
        type: string
      pagination: {}
      status:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
  description: This is the e-invoicing service API documentation.
  termsOfService: http://swagger.io/terms/
  title: Gention E-invoicing Service API
  version: "1.0"
paths:
...
  /invoice/generate-irn:
    post:
      consumes:
      - application/json
      description: Generates a new IRN for an invoice.
      parameters:
      - description: Generate IRN Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/firs_models.GenerateIRNRequestData'
      produces:
      - application/json
      responses:
        "200":
          description: IRN generated successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Generate IRN
      tags:
      - Invoice
  /invoice/sign:
    post:
      consumes:
      - application/json
      description: Signs an invoice and generates a digital signature.
      parameters:
      - description: Invoice Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/firs_models.InvoiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Invoice signed successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Sign Invoice
      tags:
      - Invoice
  /invoice/sign-irn:
    post:
      consumes:
      - application/json
      description: Signs an IRN and generates a QR code.
      parameters:
      - description: IRN Signing Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/firs_models.IRNSigningRequestData'
      produces:
      - application/json
      responses:
        "200":
          description: IRN signed successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Sign IRN
      tags:
      - Invoice
  /invoice/transmit/{irn}:
    post:
      consumes:
      - application/json
      description: Transmits an invoice to FIRS using the IRN.
      parameters:
      - description: Invoice Reference Number (IRN)
        in: path
        name: irn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoice transmitted successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Transmit Invoice
      tags:
      - Invoice
  /invoice/transmit/confirm/{irn}:
    get:
      consumes:
      - application/json
      description: Confirms a transmitted invoice using the IRN.
      parameters:
      - description: Invoice Reference Number (IRN)
        in: path
        name: irn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoice confirmed successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Confirm Transmitted Invoice
      tags:
      - Invoice
...
  /invoice/transmit/lookup-irn/{irn}:
    get:
      consumes:
      - application/json
      description: Retrieves invoice details using the IRN.
      parameters:
      - description: Invoice Reference Number (IRN)
        in: path
        name: irn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoice details retrieved
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Look Up IRN
      tags:
      - Invoice
  /invoice/transmit/lookup-party/{party_id}:
    get:
      consumes:
      - application/json
      description: Retrieves details using Party ID.
      parameters:
      - description: Party ID
        in: path
        name: party_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Party ID details retrieved
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Look Up Party ID
      tags:
      - Invoice
  /invoice/transmit/lookup-tin/{tin}:
    get:
      consumes:
      - application/json
      description: Retrieves taxpayer details using TIN.
      parameters:
      - description: Tax Identification Number (TIN)
        in: path
        name: tin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: TIN details retrieved
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Look Up TIN
      tags:
      - Invoice
  /invoice/transmit/pull:
    get:
      consumes:
      - application/json
      description: Pulls invoices from FIRS using query params.
      parameters:
      - in: query
        name: confirmed
        type: string
      - in: query
        name: from
        type: string
      - in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoices pulled successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Pull Transmitted Invoices
      tags:
      - Invoice
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"